GBAY-114 Changed int to int64

+++ DataAccess/Interfaces/ICommercialIdRepository.cs
@@ -4,11 +4,11 @@ namespace dotnet.DataAccess.Interfaces
 {
 	public interface ICommercialIdRepository
 	{
-		Task<CommercialId> GetByIdAsync(int id);
+		Task<CommercialId> GetByIdAsync(Int64 id);
 		Task<IEnumerable<CommercialId>> GetAllAsync();
 		Task AddAsync(CommercialId commercialId);
 		Task UpdateAsync(CommercialId commercialId);
-		Task DeleteAsync(int id);
+		Task DeleteAsync(Int64 id);
 
 		// Additional get methods
 		Task<IEnumerable<CommercialId>> GetByNameAsync(string name);
